KERNEL :: Algoritmo de Planificacion VRR implementado, manejo de recursos WAIT_KERNEL y SIGNAL_KERNEL implementadas, OPCION proceso estado modificada para permitir mostrar procesos bloqueados por recursos, OPCION detener planificacion implementada, OPCION iniciar planificacion implementada  CPU :: Peticiones WAIT y SIGNAL implementadas. ENTRADASALIDA :: Implementacion de bitmap, bloques y archivos para file system con asignacion contigua DIALFS, peticiones de creacion de archivos, truncamiento de archivos y eliminacion de archivos implementadas, compactacion de archivos implementada. UTILS :: nuevo tipo de dato serializable (int32_t)

+++ utils/src/comunicacion/comunicacion.h
@@ -16,7 +16,7 @@
 ///////////////////////////////////////// WRAPPERS //////////////////////////////////////
 void avisar_nuevo_proceso_memoria(int conexion_memoria, int pid, char* path_intrucciones);
 void avisar_desalojo_a_cpu(int conexion_cpu, int op_code, char* motivo);
-void solicitar_intruccion_a_memoria(int conexion, int op_code, int pid, int pc);
+void solicitar_intruccion_a_memoria(int conexion, uint8_t op_code, uint32_t pid, uint32_t pc);
 /////////////////////////////////////////////////////////////////////////////////////////
 
 // - envio generico de un op_code
